% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{cc_spaces_fg}
\alias{cc_spaces_fg}
\title{Calculate Family Group Child Care Spaces}
\usage{
cc_spaces_fg(rooms)
}
\arguments{
\item{rooms}{Number of rooms allocated for family group care.}
}
\value{
Estimated number of family group child care spaces.
}
\description{
Multiplies the number of rooms by 15 to estimate available family group child care spaces.
}
