% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{cc_spaces_infant}
\alias{cc_spaces_infant}
\title{Calculate Infant Child Care Spaces}
\usage{
cc_spaces_infant(rooms)
}
\arguments{
\item{rooms}{Number of rooms allocated for infant care.}
}
\value{
Estimated number of infant child care spaces.
}
\description{
Multiplies the number of rooms by 10 to estimate available infant child care spaces.
}
