% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{cc_spaces_preschool}
\alias{cc_spaces_preschool}
\title{Calculate Preschool Child Care Spaces}
\usage{
cc_spaces_preschool(rooms)
}
\arguments{
\item{rooms}{Number of rooms allocated for preschool care.}
}
\value{
Estimated number of preschool child care spaces.
}
\description{
Multiplies the number of rooms by 24 to estimate available preschool child care spaces.
}
