% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{cc_spaces_toddler}
\alias{cc_spaces_toddler}
\title{Calculate Toddler Child Care Spaces}
\usage{
cc_spaces_toddler(rooms)
}
\arguments{
\item{rooms}{Number of rooms allocated for toddler care.}
}
\value{
Estimated number of toddler child care spaces.
}
\description{
Multiplies the number of rooms by 15 to estimate available toddler child care spaces.
}
