% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{fiscal_year}
\alias{fiscal_year}
\title{Determine Fiscal Year}
\usage{
fiscal_year(date)
}
\arguments{
\item{date}{A date object or a string that can be coerced to a date.}
}
\value{
A character string representing the fiscal year, in the format "YYYY-YY".
If the input date is invalid or `NA`, the function returns `NA_character_`.
}
\description{
This function takes a date and returns a string representing the fiscal year in the format "YYYY-YY".
The fiscal year is assumed to run from April of one year to March of the next year.
}
\examples{
fiscal_year("2024-03-31")
# Returns: "2023-24"

fiscal_year("2023-04-01")
# Returns: "2023-24"

fiscal_year(NA)
# Returns: NA_character_

}
