% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{frank_columns}
\alias{frank_columns}
\title{Original Frankenstein column names}
\usage{
frank_columns()
}
\value{
A character vector of the original Frankenstein column names.
}
\description{
Original Frankenstein column names
}
