% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{monthly_all}
\alias{monthly_all}
\title{Retrieve Monthly All Data}
\usage{
monthly_all(
  as_of = NULL,
  everything = FALSE,
  schema = NULL,
  database = NULL,
  address = NULL
)
}
\arguments{
\item{as_of}{The date of the Monthly All data to be pulled.}

\item{everything}{If FALSE (default), intended to return only a default column set once
one is defined. Currently has no effect.}

\item{schema}{The schema name (default: NULL, will use `ezql_details_schema`).}

\item{database}{The database name (default: NULL, will use `ezql_details_db`).}

\item{address}{The server address (default: NULL, will use `ezql_details_add`).}
}
\value{
A tibble containing the data from the specified SQL table.
}
\description{
This function retrieves data from the specified SQL table.
}
