% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pt}
\alias{pt}
\title{Convert Event-Level Data to Project-Level Data}
\usage{
pt(sql_table_name = "Events", schema = NULL, database = NULL, address = NULL)
}
\arguments{
\item{sql_table_name}{The name of the SQL table (default: "Events").}

\item{schema}{The schema name (default: NULL, will use `ezql_details_schema`).}

\item{database}{The database name (default: NULL, will use `ezql_details_db`).}

\item{address}{The server address (default: NULL, will use `ezql_details_add`).}
}
\value{
A tibble containing the project-level data.
}
\description{
This function takes event-level data and converts it to project-level data by summarizing
specific columns and filling down other columns.
}
