% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pt_columns}
\alias{pt_columns}
\title{Column set for project-level (pt) output}
\usage{
pt_columns()
}
\value{
A character vector of column names.
}
\description{
Returns the canonical column order for project-level summaries.
}
\examples{
\dontrun{
pt_columns()
}
}
