% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{school_year}
\alias{school_year}
\title{Determine School Year}
\usage{
school_year(date)
}
\arguments{
\item{date}{A date object or a string that can be coerced to a date.}
}
\value{
A character string representing the school year, in the format "YYYY-YY".
If the input date is invalid or `NA`, the function returns `NA_character_`.
}
\description{
This function takes a date and returns a string representing the school year in the format "YYYY-YY".
The school year runs from September of one year to August of the next year.
}
\examples{
school_year("2024-05-15")
# Returns: "2023-24"

school_year("2023-09-01")
# Returns: "2023-24"

school_year(NA)
# Returns: NA_character_

}
