% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{frank}
\alias{frank}
\title{Retrieve Frankenstein Data}
\usage{
frank(
  sql_table_name = "Frankenstein",
  schema = NULL,
  database = NULL,
  address = NULL
)
}
\arguments{
\item{sql_table_name}{The name of the SQL table (default: "Frankenstein").}

\item{schema}{The schema name (default: NULL, will use `ezql_details_schema`).}

\item{database}{The database name (default: NULL, will use `ezql_details_db`).}

\item{address}{The server address (default: NULL, will use `ezql_details_add`).}
}
\value{
A tibble containing the data from the specified SQL table.
}
\description{
This function retrieves data from the specified SQL table.
}
